+++ packages/google_mobile_ads/test/ump/consent_form_test.dart
@@ -24,7 +24,7 @@ import 'package:mockito/mockito.dart';
 
 import 'consent_form_test.mocks.dart';
 
-@GenerateMocks([UserMessagingChannel])
+@GenerateNiceMocks(<MockSpec<Object>>[MockSpec<UserMessagingChannel>()])
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
@@ -34,6 +34,29 @@ void main() {
     setUp(() async {
       mockChannel = MockUserMessagingChannel();
       UserMessagingChannel.instance = mockChannel;
+    });
+
+    test('showPrivacyOptionsForm success', () async {
+      Completer<FormError?> formCompleter = Completer<FormError?>();
+
+      await ConsentForm.showPrivacyOptionsForm(
+          (formError) => formCompleter.complete(formError));
+
+      FormError? formError = await formCompleter.future;
+      expect(formError, null);
+    });
+
+    test('showPrivacyOptionsForm failure', () async {
+      FormError? testError = FormError(errorCode: 1, message: 'testErrorMsg');
+      when(mockChannel.showPrivacyOptionsForm())
+          .thenAnswer((realInvocation) => Future<FormError?>.value(testError));
+      Completer<FormError?> formCompleter = Completer<FormError?>();
+
+      await ConsentForm.showPrivacyOptionsForm(
+          (formError) => formCompleter.complete(formError));
+
+      FormError? formError = await formCompleter.future;
+      expect(formError, testError);
     });
 
     test('loadConsentForm() success', () async {
@@ -72,6 +95,29 @@ void main() {
       expect(formError, responseError);
       expect(successCompleter.isCompleted, false);
       expect(errorCompleter.isCompleted, true);
+    });
+
+    test('loadAndShowConsentFormIfRequired success', () async {
+      Completer<FormError?> formCompleter = Completer<FormError?>();
+
+      await ConsentForm.loadAndShowConsentFormIfRequired(
+          (formError) => formCompleter.complete(formError));
+
+      FormError? formError = await formCompleter.future;
+      expect(formError, null);
+    });
+
+    test('loadAndShowConsentFormIfRequired failure', () async {
+      FormError? testError = FormError(errorCode: 1, message: 'testErrorMsg');
+      when(mockChannel.loadAndShowConsentFormIfRequired())
+          .thenAnswer((realInvocation) => Future<FormError?>.value(testError));
+      Completer<FormError?> formCompleter = Completer<FormError?>();
+
+      await ConsentForm.loadAndShowConsentFormIfRequired(
+          (formError) => formCompleter.complete(formError));
+
+      FormError? formError = await formCompleter.future;
+      expect(formError, testError);
     });
   });
 }