+++ packages/mediation/gma_mediation_pangle/example/android/app/src/main/kotlin/io/flutter/plugins/googlemobileads/mediation/gma_mediation_applovin_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package io.flutter.plugins.googlemobileads.mediation.gma_mediation_pangle_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()