+++ packages/mediation/gma_mediation_meta/CHANGELOG.md
@@ -1,5 +1,15 @@
-## 1.0.0
+## Meta Audience Network Flutter Mediation Adapter Changelog
+
+#### 1.0.1
+
+* Added dart class to allow correct compatibility analysis with host platforms
+* Verified compatibility with Meta Android adapter version 6.17.0.0
+* Verified compatibility with Meta iOS adapter version 6.15.0.0
+* Built and tested with the Google Mobile Ads Flutter Plugin version 5.1.0.
+
+#### 1.0.0
 
 * Initial release.
 * Verified compatibility with Meta Android adapter version 6.17.0.0 
 * Verified compatibility with Meta iOS adapter version 6.15.0.0
+* Built and tested with the Google Mobile Ads Flutter Plugin version 5.0.0.